% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{GBAccession-class}
\alias{GBAccession}
\alias{GBAccession-class}
\title{GBAccession ID class}
\usage{
GBAccession(id)
}
\arguments{
\item{id}{A versioned GenBank Accession id}
}
\value{
a \code{GBAccession} object.
}
\description{
A class representing the (versioned) GenBank accession
}
\examples{
id = GBAccession("U49845.1")
\dontrun{gb = readGenBank(id)}
}

