% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{gbk-specific-api}
\alias{accession}
\alias{accession,GenBankAnnot}
\alias{accession,GenBankAnnot-method}
\alias{accession,GenBankFull}
\alias{accession,GenBankFull-method}
\alias{gbk-specific-api}
\alias{vers}
\alias{vers,GenBankAnnot}
\alias{vers,GenBankAnnot-method}
\alias{vers,GenBankFull}
\alias{vers,GenBankFull-method}
\title{genbankr specific api}
\usage{
accession(x, ...)

\S4method{accession}{GenBankAnnot}(x)

\S4method{accession}{GenBankFull}(x)

vers(x, ...)

\S4method{vers}{GenBankAnnot}(x)

\S4method{vers}{GenBankFull}(x)
}
\arguments{
\item{x}{A genbank annotation object}

\item{...}{unused.}
}
\value{
Character vectors for \code{accession} and \code{vers}
}
\description{
Accessor functions specific to genbankr objects.
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
accession(gb)
vers(gb)
}

