% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/api-methods.R
\docType{methods}
\name{GenBankFile-class}
\alias{GBFile-class}
\alias{GBKFile-class}
\alias{GenBankFile}
\alias{GenBankFile-class}
\title{GenBank File}
\usage{
GenBankFile(fil)
}
\arguments{
\item{fil}{character. Path to the genbank file}
}
\value{
A \code{GenBankFile} object
}
\description{
A resource class for use within the rtracklayer framework

Create a GenBankFile object.
}
\examples{
fil = GenBankFile(system.file("sample.gbk", package="genbankr"))
gb = import(fil)
}

