% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{GenBankAnnot-class}
\alias{GenBankAnnot-class}
\alias{GenBankFull-class}
\title{GenBank data objects}
\description{
These objects represent GenBank annotations (GenBankAnnot) and
those annotations paired with raw reference/origin sequence
(GenBankFull).
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
gb
}

