% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{import,GenBankFile,ANY,ANY-method}
\alias{import,GenBankFile}
\alias{import,GenBankFile,ANY,ANY-method}
\title{Import genbank file}
\usage{
\S4method{import}{GenBankFile,ANY,ANY}(con, format, text, ...)
}
\arguments{
\item{con}{See import docs.}

\item{format}{See import docs.}

\item{text}{See import docs.}

\item{...}{Arguments passed to \code{readGenBank}}
}
\value{
A \code{GenBankRecord} object.
}
\description{
Import a genbank file using the rtracklayer API.
}

