% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbankReader.R
\name{make_gbrecord}
\alias{make_gbrecord}
\title{GenBank object constructors}
\usage{
make_gbrecord(rawgbk, verbose = FALSE)
}
\arguments{
\item{rawgbk}{list. The output of \code{parseGenBank}}

\item{verbose}{logical. Should informative messages be shown}
}
\value{
A GenBankRecord object
}
\description{
Constructors for \code{GenBankRecord} objects.
}
\examples{
prsed = parseGenBank(system.file("sample.gbk", package="genbankr"))
gb = make_gbrecord(prsed)
}

