% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{variants}
\alias{variants}
\alias{variants,GenBankRecord}
\alias{variants,GenBankRecord-method}
\title{Retrieve variantion features}
\usage{
variants(x)

\S4method{variants}{GenBankRecord}(x)
}
\arguments{
\item{x}{a GenBankRecord object}
}
\value{
A VRanges containing the variations annotated in the source file
}
\description{
Extract the annotated variants from a GenBankRecord object
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
variants(gb)
}

