% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{cds,GenBankAnnot-method}
\alias{annotations}
\alias{annotations,GenBankFull-method}
\alias{cds,GenBankAnnot-method}
\alias{cds,GenBankFull-method}
\alias{cdsBy,GenBankAnnot}
\alias{cdsBy,GenBankAnnot-method}
\alias{cdsBy,GenBankFull}
\alias{cdsBy,GenBankFull-method}
\alias{exons,GenBankAnnot-method}
\alias{exons,GenBankFull-method}
\alias{exonsBy,GenBankAnnot}
\alias{exonsBy,GenBankAnnot-method}
\alias{exonsBy,GenBankFull}
\alias{exonsBy,GenBankFull-method}
\alias{genes,GenBankAnnot-method}
\alias{genes,GenBankFull-method}
\alias{getSeq,GBAccession-method}
\alias{getSeq,GenBankAnnot-method}
\alias{getSeq,GenBankFile-method}
\alias{getSeq,GenBankFull-method}
\alias{isCircular,GenBankAnnot}
\alias{isCircular,GenBankAnnot-method}
\alias{isCircular,GenBankFull}
\alias{isCircular,GenBankFull-method}
\alias{seqinfo,GenBankAnnot}
\alias{seqinfo,GenBankAnnot-method}
\alias{seqinfo,GenBankFull}
\alias{seqinfo,GenBankFull-method}
\alias{transcripts,GenBankAnnot-method}
\alias{transcripts,GenBankFull-method}
\title{Annotation extraction api}
\usage{
\S4method{cds}{GenBankAnnot}(x)

\S4method{exons}{GenBankAnnot}(x)

\S4method{genes}{GenBankAnnot}(x)

\S4method{transcripts}{GenBankAnnot}(x)

\S4method{cds}{GenBankFull}(x)

\S4method{exons}{GenBankFull}(x)

\S4method{genes}{GenBankFull}(x)

\S4method{transcripts}{GenBankFull}(x)

annotations(x)

\S4method{annotations}{GenBankFull}(x)

\S4method{getSeq}{GenBankFull}(x, ...)

\S4method{getSeq}{GenBankAnnot}(x, ...)

\S4method{getSeq}{GenBankFile}(x, ...)

\S4method{getSeq}{GBAccession}(x, ...)

\S4method{cdsBy}{GenBankAnnot}(x, by = c("tx", "gene"))

\S4method{exonsBy}{GenBankAnnot}(x, by = c("tx", "gene"))

\S4method{cdsBy}{GenBankFull}(x, by = c("tx", "gene"))

\S4method{exonsBy}{GenBankFull}(x, by = c("tx", "gene"))

\S4method{isCircular}{GenBankAnnot}(x)

\S4method{isCircular}{GenBankFull}(x)

\S4method{seqinfo}{GenBankAnnot}(x)

\S4method{seqinfo}{GenBankFull}(x)
}
\arguments{
\item{x}{The object containing the annotations}

\item{...}{unused.}

\item{by}{character. Factor to group the resulting GRanges by.}
}
\value{
The expected types, \code{GenomicRanges} for most functions,
a \code{DNAStrimgSet} for \code{getSeq}, a \code{GenBankAnnot} for
\code{annotations}.
}
\description{
Accessor functions shared with the larger Bioconductor
ecosystem.
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
cds(gb)
exons(gb)
genes(gb)
}

