% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{gbk-specific-api}
\alias{accession}
\alias{accession,GenBankRecord}
\alias{accession,GenBankRecord-method}
\alias{definition}
\alias{definition,GenBankRecord}
\alias{definition,GenBankRecord-method}
\alias{gbk-specific-api}
\alias{locus}
\alias{locus,GenBankRecord}
\alias{locus,GenBankRecord-method}
\alias{sources}
\alias{sources,GenBankRecord}
\alias{sources,GenBankRecord-method}
\alias{vers}
\alias{vers,GenBankRecord}
\alias{vers,GenBankRecord-method}
\title{genbankr specific api}
\usage{
accession(x, ...)

\S4method{accession}{GenBankRecord}(x)

definition(x, ...)

\S4method{definition}{GenBankRecord}(x)

locus(x, ...)

\S4method{locus}{GenBankRecord}(x)

vers(x, ...)

\S4method{vers}{GenBankRecord}(x)

sources(x, ...)

\S4method{sources}{GenBankRecord}(x)
}
\arguments{
\item{x}{A genbank annotation object}

\item{...}{unused.}
}
\value{
Character vectors for \code{accession} and \code{vers}
}
\description{
Accessor functions specific to genbankr objects.
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
accession(gb)
vers(gb)
}

