% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{GenBankRecord-class}
\alias{GenBankRecord-class}
\title{GenBank data objects}
\description{
These objects represent GenBank annotations
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
gb
}

