% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{intergenic}
\alias{intergenic}
\alias{intergenic,GenBankRecord-method}
\title{Extract intergenic regions from processed GenBank annotations}
\usage{
\S4method{intergenic}{GenBankRecord}(x)
}
\arguments{
\item{x}{A GenBankRecord object}
}
\value{
A GRanges for the intergenic regions, defined as regions not
overlapping any genes defined in the annotations on either strand.
}
\description{
Extract the intergenic regions from a set of GenBank annotations.
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
intergenic(gb)
}

