% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbankReader.R
\name{make_gbannot}
\alias{make_gbannot}
\alias{make_gbfull}
\title{GenBank object constructors}
\usage{
make_gbannot(rawgbk, verbose = FALSE)

make_gbfull(rawgbk, verbose = FALSE)
}
\arguments{
\item{rawgbk}{list. The output of \code{parseGenBank}}

\item{verbose}{logical. Should informative messages be shown}
}
\value{
A GenBankAnnot (\code{make_gbannot}) or
GenBankFull (\code{make_gbfull}) object.
}
\description{
Constructors for \code{GenBankFull} and
\code{GenBankAnnot} objects.
}
\examples{
prsed = parseGenBank(system.file("sample.gbk", package="genbankr"))
gb = make_gbannot(prsed)
}

