% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{otherFeatures}
\alias{otherFeatures}
\alias{otherFeatures,GenBankRecord}
\alias{otherFeatures,GenBankRecord-method}
\title{Retrieve 'other' features}
\usage{
otherFeatures(x)

\S4method{otherFeatures}{GenBankRecord}(x)
}
\arguments{
\item{x}{a GenBankRecord object}
}
\value{
A GRanges containing the features which don't fall into another
category (ie not gene, exon, transcript, cds, or variant) annotated in the
source file
}
\description{
Retrieve  the other features (not covered by a different accessor)
from the set of annotations
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
otherFeatures(gb)
}

