% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbankReader.R
\name{parseGenBank}
\alias{parseGenBank}
\title{Parse raw genbank file content}
\usage{
parseGenBank(file, text = readLines(file), partial = NA, verbose = FALSE,
  ret.anno = TRUE, ret.seq = TRUE)
}
\arguments{
\item{file}{character. The file to be parsed. Ignored if \code{text} is
specified}

\item{text}{character. The text to be parsed.}

\item{partial}{logical. If TRUE, features with non-exact boundaries will
be included. Otherwise, non-exact features are excluded, with a warning
if \code{partial} is \code{NA} (the default).}

\item{verbose}{logical. Should informative messages be printed to the
console as the file is being processed.}

\item{ret.anno}{logical. Should the annotations in the GenBank file be
parsed and included in the returned object. (Defaults to \code{TRUE})}

\item{ret.seq}{logical. Should the origin sequence (if present) in the
GenBank file be included in the returned object. (Defaults to \code{TRUE})}
}
\value{
if \code{ret.anno} is \code{TRUE}, a list containing the parsed
contents of the file, suitable for passing to \code{make_gbrecord}. If
\code{ret.anno} is \code{FALSE}, a \code{DNAStringSet} object containing
the origin sequence.
}
\description{
Parse genbank content and return a low-level list object
containing each component of the file.
}
\note{
This is a low level function not intended for common end-user use.
In nearly all cases, end-users (and most developers) should call
\code{readGenBank} or create a \code{GenBankFile} object and call
\code{import} instead.
}
\examples{
prsd = parseGenBank(system.file("sample.gbk", package="genbankr"))
}

