% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-methods.R
\docType{methods}
\name{variants}
\alias{variants}
\alias{variants,GenBankAnnot}
\alias{variants,GenBankAnnot-method}
\alias{variants,GenBankFull}
\alias{variants,GenBankFull-method}
\title{Retrieve variantion features}
\usage{
variants(x)

\S4method{variants}{GenBankAnnot}(x)

\S4method{variants}{GenBankFull}(x)
}
\arguments{
\item{x}{a GenBankAnnot or GenBankFull object}
}
\value{
A VRanges containing the variations annotated in the source file
}
\description{
Extract the annotated variants from a GenBankAnnot or GenBankFull object
}
\examples{
gb = readGenBank(system.file("sample.gbk", package="genbankr"))
variants(gb)
}

